VISTA-56 Change All Float to Double

+++ vista-portal/vista-portal-domain/src/main/java/com/propertyvista/portal/domain/PropertyOwner.java
@@ -36,7 +36,7 @@ public interface PropertyOwner extends IEntity {
     /**
      * Percent of share
      */
-    IPrimitive<Float> share();
+    IPrimitive<Double> share();
 
     IPrimitive<Date> startDate();
