+++ vista-portal/vista-portal-domain/src/main/java/com/propertyvista/portal/domain/pt/Address.java
@@ -46,7 +46,7 @@ public interface Address extends IEntity {
 
     IPrimitive<Date> moveOutDate();
 
-    IPrimitive<Float> payment();
+    IPrimitive<Double> payment();
 
     IPrimitive<String> phone();
