+++ vista-portal/vista-portal-domain/src/main/java/com/propertyvista/portal/domain/pt/PotentialTenant.java
@@ -49,7 +49,7 @@ public interface PotentialTenant extends IApplicationEntity {
 
     IPrimitive<Relationship> relationship();
 
-    IPrimitive<Float> payment();
+    IPrimitive<Double> payment();
 
     IPrimitive<Boolean> dependant();
