+++ vista-portal/vista-portal-domain/src/main/java/com/propertyvista/portal/domain/pt/TenantAsset.java
@@ -23,5 +23,5 @@ public interface TenantAsset extends IEntity {
     IPrimitive<String> name();
 
     @Caption(name = "Value")
-    IPrimitive<Float> assetValue();
+    IPrimitive<Double> assetValue();
 }
